ffi: migrate cstrFromString / read* / write* family to @intrinsic in libs/std/ffi.il
Previously the FFI marshalling helpers (cstrFromString, stringFromCstr,
freeCstr, cstrArrayToStrings, bytesFromBuffer, errnoCheck/I64, readT/writeT
for every primitive width) were compiler-magic bare names — the type
checker had a hardcoded FFI_HELPERS list and MIR call lowering intercepted
each name to emit a direct runtime call. The shape was opaque to users
and to the LSP, and the helpers leaked into top-level namespace.

Move them into libs/std/ffi.il as ordinary `@intrinsic("ffi.X")` fn
declarations inside one `@extern(C) { ... }` block. User code reaches
them through `use std.ffi { ... }`, and the regular @intrinsic
declaration path handles MIR lowering / cranelift import.

read*/write* keep narrow argument / return types (i8, u32, f64, ...)
matching the runtime signatures in raw_mem.rs.

The @objc desugar still synthesises bare cstrFromString calls during
parsing, before any `use` resolution runs. Route those through
`\$ffi.cstrFromString` instead — the `\$` prefix is unreachable from user
code (lex rejects it), so the name is unambiguously parser-synthesised
and bypasses the use-import machinery via a small type-check / lower
special-case plus a cranelift Import in program_decl.

bytesFromBuffer's signature is `(\*const void, size_t)` to match the
typical caller pattern.

Touched 68 .il consumer files to add `use std.ffi { ... }` for the
helpers they call; cleared the now-stale fixture that asserted the
old "errnoCheck must be inside @extern(C)" rule (errnoCheck takes /
returns ordinary ilang types, so the rule no longer applies).

+++ crates/ilang-runtime/src/raw_mem.rs
@@ -1,38 +1,43 @@
 //! Raw memory FFI read / write helpers. Mostly used by `@extern(C)`
 //! call sites that need to poke specific byte sizes into a C-side
 //! struct buffer.
+//!
+//! Each helper's signature mirrors the ilang `@intrinsic` declaration
+//! in `libs/std/ffi.il` exactly — narrow integer reads return the
+//! actual integer width (not i64), so the cranelift ABI on either
+//! side agrees byte-for-byte. Writes take the narrow value, too.
 
 #[unsafe(export_name = "$ffi.readI8")]
-pub extern "C" fn __read_i8(p: i64, off: i64) -> i64 {
-    unsafe { *((p + off) as *const i8) as i64 }
+pub extern "C" fn __read_i8(p: i64, off: i64) -> i8 {
+    unsafe { *((p + off) as *const i8) }
 }
 #[unsafe(export_name = "$ffi.readI16")]
-pub extern "C" fn __read_i16(p: i64, off: i64) -> i64 {
-    unsafe { *((p + off) as *const i16) as i64 }
+pub extern "C" fn __read_i16(p: i64, off: i64) -> i16 {
+    unsafe { *((p + off) as *const i16) }
 }
 #[unsafe(export_name = "$ffi.readI32")]
-pub extern "C" fn __read_i32(p: i64, off: i64) -> i64 {
-    unsafe { *((p + off) as *const i32) as i64 }
+pub extern "C" fn __read_i32(p: i64, off: i64) -> i32 {
+    unsafe { *((p + off) as *const i32) }
 }
 #[unsafe(export_name = "$ffi.readI64")]
 pub extern "C" fn __read_i64(p: i64, off: i64) -> i64 {
     unsafe { *((p + off) as *const i64) }
 }
 #[unsafe(export_name = "$ffi.readU8")]
-pub extern "C" fn __read_u8(p: i64, off: i64) -> i64 {
-    unsafe { *((p + off) as *const u8) as i64 }
+pub extern "C" fn __read_u8(p: i64, off: i64) -> u8 {
+    unsafe { *((p + off) as *const u8) }
 }
 #[unsafe(export_name = "$ffi.readU16")]
-pub extern "C" fn __read_u16(p: i64, off: i64) -> i64 {
-    unsafe { *((p + off) as *const u16) as i64 }
+pub extern "C" fn __read_u16(p: i64, off: i64) -> u16 {
+    unsafe { *((p + off) as *const u16) }
 }
 #[unsafe(export_name = "$ffi.readU32")]
-pub extern "C" fn __read_u32(p: i64, off: i64) -> i64 {
-    unsafe { *((p + off) as *const u32) as i64 }
+pub extern "C" fn __read_u32(p: i64, off: i64) -> u32 {
+    unsafe { *((p + off) as *const u32) }
 }
 #[unsafe(export_name = "$ffi.readU64")]
-pub extern "C" fn __read_u64(p: i64, off: i64) -> i64 {
-    unsafe { *((p + off) as *const u64) as i64 }
+pub extern "C" fn __read_u64(p: i64, off: i64) -> u64 {
+    unsafe { *((p + off) as *const u64) }
 }
 #[unsafe(export_name = "$ffi.readF32")]
 pub extern "C" fn __read_f32(p: i64, off: i64) -> f32 {
@@ -44,36 +49,36 @@ pub extern "C" fn __read_f64(p: i64, off: i64) -> f64 {
 }
 
 #[unsafe(export_name = "$ffi.writeI8")]
-pub extern "C" fn __write_i8(p: i64, off: i64, v: i64) {
-    unsafe { *((p + off) as *mut i8) = v as i8; }
+pub extern "C" fn __write_i8(p: i64, off: i64, v: i8) {
+    unsafe { *((p + off) as *mut i8) = v; }
 }
 #[unsafe(export_name = "$ffi.writeI16")]
-pub extern "C" fn __write_i16(p: i64, off: i64, v: i64) {
-    unsafe { *((p + off) as *mut i16) = v as i16; }
+pub extern "C" fn __write_i16(p: i64, off: i64, v: i16) {
+    unsafe { *((p + off) as *mut i16) = v; }
 }
 #[unsafe(export_name = "$ffi.writeI32")]
-pub extern "C" fn __write_i32(p: i64, off: i64, v: i64) {
-    unsafe { *((p + off) as *mut i32) = v as i32; }
+pub extern "C" fn __write_i32(p: i64, off: i64, v: i32) {
+    unsafe { *((p + off) as *mut i32) = v; }
 }
 #[unsafe(export_name = "$ffi.writeI64")]
 pub extern "C" fn __write_i64(p: i64, off: i64, v: i64) {
     unsafe { *((p + off) as *mut i64) = v; }
 }
 #[unsafe(export_name = "$ffi.writeU8")]
-pub extern "C" fn __write_u8(p: i64, off: i64, v: i64) {
-    unsafe { *((p + off) as *mut u8) = v as u8; }
+pub extern "C" fn __write_u8(p: i64, off: i64, v: u8) {
+    unsafe { *((p + off) as *mut u8) = v; }
 }
 #[unsafe(export_name = "$ffi.writeU16")]
-pub extern "C" fn __write_u16(p: i64, off: i64, v: i64) {
-    unsafe { *((p + off) as *mut u16) = v as u16; }
+pub extern "C" fn __write_u16(p: i64, off: i64, v: u16) {
+    unsafe { *((p + off) as *mut u16) = v; }
 }
 #[unsafe(export_name = "$ffi.writeU32")]
-pub extern "C" fn __write_u32(p: i64, off: i64, v: i64) {
-    unsafe { *((p + off) as *mut u32) = v as u32; }
+pub extern "C" fn __write_u32(p: i64, off: i64, v: u32) {
+    unsafe { *((p + off) as *mut u32) = v; }
 }
 #[unsafe(export_name = "$ffi.writeU64")]
-pub extern "C" fn __write_u64(p: i64, off: i64, v: i64) {
-    unsafe { *((p + off) as *mut u64) = v as u64; }
+pub extern "C" fn __write_u64(p: i64, off: i64, v: u64) {
+    unsafe { *((p + off) as *mut u64) = v; }
 }
 #[unsafe(export_name = "$ffi.writeF32")]
 pub extern "C" fn __write_f32(p: i64, off: i64, v: f32) {
